[src] Move functions for collator
